Adds configuration file support
Introduces support for a configuration file (`act.xml`) to define paths for steps and scenarios.

This allows users to customize the location of these files, making the framework more flexible.

Adds command line options `-c` and `--configuration` to specify a configuration file.

If no configuration file is specified, it defaults to `act.xml`.
If the config file is missing, the paths default to `tests/steps` and `tests/scenarios`.